+++ packages/lib/bookings/hostGroupUtils.ts
@@ -37,3 +37,24 @@ export function getHostsFromOtherGroups<T extends { groupId?: string | null }>(
     (host) => (groupId && (!host.groupId || host.groupId !== groupId)) || (!groupId && host.groupId)
   );
 }
+
+export function sortHosts(
+  hostA: { priority: number | null; weight: number | null },
+  hostB: { priority: number | null; weight: number | null },
+  isRRWeightsEnabled: boolean
+) {
+  const weightA = hostA.weight ?? 100;
+  const priorityA = hostA.priority ?? 2;
+  const weightB = hostB.weight ?? 100;
+  const priorityB = hostB.priority ?? 2;
+
+  if (isRRWeightsEnabled) {
+    if (weightA === weightB) {
+      return priorityB - priorityA;
+    } else {
+      return weightB - weightA;
+    }
+  } else {
+    return priorityB - priorityA;
+  }
+}
