+++ packages/platform/atoms/package.json
@@ -63,6 +63,9 @@
     "./dist/index.ts": "./index.ts",
     "./dist/index.d.ts": "./dist/index.d.ts",
     "./event-types/wrappers/EventTypeWebWrapper": "./event-types/wrappers/EventTypeWebWrapper.tsx",
+    "./event-types/hooks/useEventTypeForm": "./event-types/hooks/useEventTypeForm.ts",
+    "./event-types/hooks/useHandleRouteChange": "./event-types/hooks/useHandleRouteChange.ts",
+    "./event-types/hooks/useTabsNavigations": "./event-types/hooks/useTabsNavigations.tsx",
     "./globals.min.css": "./globals.min.css",
     "./hooks/bookings/useBookings": "./hooks/bookings/useBookings.ts",
     "./hooks/bookings/useHandleBookEvent": "./hooks/bookings/useHandleBookEvent.ts",
